refactor(hackathon): 해커톤 타입·데이터를 @repo/components 공통 패키지로 이동
