+++ packages/components/src/config/index.ts
@@ -1,1 +1,6 @@
 export { PROJECT_TYPE_LABELS, PROJECT_TYPE_STYLES } from "./projectTypes";
+export {
+  HACKATHONS_DATA,
+  HACKATHON_CATEGORY_STYLES,
+} from "./hackathonTypes";
+export type { HackathonData, HackathonCategory } from "./hackathonTypes";
